+++ app/src/app/views/stix/group/group-view/group-view.component.ts
@@ -3,7 +3,7 @@ import { Group } from 'src/app/classes/stix/group';
 import { StixViewPage } from '../../stix-view-page';
 import { Relationship } from 'src/app/classes/stix/relationship';
 import { AuthenticationService } from 'src/app/services/connectors/authentication/authentication.service';
-import { RestApiConnectorService } from "src/app/services/connectors/rest-api/rest-api-connector.service";
+import { RestApiConnectorService } from 'src/app/services/connectors/rest-api/rest-api-connector.service';
 
 @Component({
     selector: 'app-group-view',
@@ -21,8 +21,8 @@ export class GroupViewComponent extends StixViewPage implements OnInit {
     }
 
     ngOnInit() {
-        if (this.group.firstInitialized) {
-            this.group.initializeWithDefaultMarkingDefinitions(this.restApiConnector);
-        }
+      if (this.group.firstInitialized) {
+        this.group.initializeWithDefaultMarkingDefinitions(this.restApiConnector);
+      }
     }
 }