+++ app/src/app/views/stix/matrix/matrix-view/matrix-view.component.ts
@@ -29,7 +29,7 @@ export class MatrixViewComponent extends StixViewPage implements OnInit {
             })
         }
         if (this.matrix.firstInitialized) {
-            this.matrix.initializeWithDefaultMarkingDefinitions(this.restAPIConnectorService)
+            this.matrix.initializeWithDefaultMarkingDefinitions(this.restAPIConnectorService);
         }
     }
 }