+++ app/src/app/views/stix/mitigation/mitigation-view/mitigation-view.component.ts
@@ -3,7 +3,7 @@ import { Mitigation } from 'src/app/classes/stix/mitigation';
 import { StixViewPage } from '../../stix-view-page';
 import { Relationship } from 'src/app/classes/stix/relationship';
 import { AuthenticationService } from 'src/app/services/connectors/authentication/authentication.service';
-import { RestApiConnectorService } from "src/app/services/connectors/rest-api/rest-api-connector.service";
+import { RestApiConnectorService } from 'src/app/services/connectors/rest-api/rest-api-connector.service';
 
 @Component({
     selector: 'app-mitigation-view',