+++ app/src/app/views/stix/stix-page/stix-page.component.ts
@@ -166,7 +166,7 @@ export class StixPageComponent implements OnInit, OnDestroy {
             })
             let subscription = prompt.afterClosed().subscribe({
                 next: (result) => {
-                    this.objects = [new Software(result)]; 
+                    this.objects = [new Software(result)];
                     this.initialVersion = new VersionNumber(this.objects[0].version.toString());
                     this.updateBreadcrumbs(this.objects, this.objectType);
                 },
@@ -176,14 +176,14 @@ export class StixPageComponent implements OnInit, OnDestroy {
             // create a new object to edit
             this.objects = [];
             this.objects.push(
-                this.objectType  == "matrix" ? new Matrix() :
-                this.objectType  == "technique" ? new Technique() :
-                this.objectType  == "tactic" ? new Tactic() :
-                this.objectType  == "mitigation" ? new Mitigation() :
-                this.objectType  == "group" ? new Group():
-                this.objectType  == "collection" ? new Collection() : 
-                this.objectType  == "data-source" ? new DataSource() :
-                this.objectType  == "marking-definition" ? new MarkingDefinition() :
+                this.objectType  == 'matrix' ? new Matrix() :
+                this.objectType  == 'technique' ? new Technique() :
+                this.objectType  == 'tactic' ? new Tactic() :
+                this.objectType  == 'mitigation' ? new Mitigation() :
+                this.objectType  == 'group' ? new Group() :
+                this.objectType  == 'collection' ? new Collection() :
+                this.objectType  == 'data-source' ? new DataSource() :
+                this.objectType  == 'marking-definition' ? new MarkingDefinition() :
                 null // if not any of the above types
             );
             this.initialVersion = new VersionNumber(this.objects[0].version.toString());